beginner/ecb_a_tron_9000/src/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pycryptodome==3.23.0
